+++ packages/demo/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [24.10.2](https://github.com/coveo/react-vapor/compare/v24.10.1...v24.10.2) (2021-09-02)
+
+**Note:** Version bump only for package react-vapor-demo
+
+
+
+
+
+## [24.10.1](https://github.com/coveo/react-vapor/compare/v24.10.0...v24.10.1) (2021-09-02)
+
+**Note:** Version bump only for package react-vapor-demo
+
+
+
+
+
+# [24.10.0](https://github.com/coveo/react-vapor/compare/v24.9.4...v24.10.0) (2021-09-01)
+
+**Note:** Version bump only for package react-vapor-demo
+
+
+
+
+
+## [24.9.4](https://github.com/coveo/react-vapor/compare/v24.9.3...v24.9.4) (2021-08-30)
+
+**Note:** Version bump only for package react-vapor-demo
+
+
+
+
+
+## [24.9.3](https://github.com/coveo/react-vapor/compare/v24.9.2...v24.9.3) (2021-08-26)
+
+**Note:** Version bump only for package react-vapor-demo
+
+
+
+
+
+## [24.9.2](https://github.com/coveo/react-vapor/compare/v24.9.1...v24.9.2) (2021-08-25)
+
+**Note:** Version bump only for package react-vapor-demo
+
+
+
+
+
 ## [24.9.1](https://github.com/coveo/react-vapor/compare/v24.9.0...v24.9.1) (2021-08-23)
 
 **Note:** Version bump only for package react-vapor-demo